move the MapPage map component into it's own directory with it's own css file
